Accept stack keyword in get_tokens_unprocessed
Code wasn't running with modern version of pygments. Adding stack
keyword to function definition fixes it.

Now returns Keyword.Type, which highlights the same as builtin types,
like "int" and "char", for the keywords "Atom", "System", and "vec3".

Also, cleaned up the code to be more similar in structure to the code
in pygments/lexers/c_cpp.py.