Reproducible build by setting PDF's date to same as git
PDF files have the date they were built embedded in them, which means
that a file like example.pdf would have a different checksum every
time pdflatex is run.

The solution is to set the SOURCE_DATE_EPOCH environment variable to
the output of `git log -1 --format=%at example.ltx`. That way
example.pdf's embedded date is close to correct: it is the author's
timestamp of the last commit of example.ltx.